fix(test): fix race conditions in Forward tests
The reason why we had race conditions is that we have a server waiting
20 ms. The proxy called this server and closed the connection after 5
ms, allowing the test to finish. Once the test was finished, the
server was still sleeping/waiting to end the request processing but
was closed by the main routine before it was able to finish
processing.